--- app/src/main/java/com/example/shopping_app/CheckoutFragment.java
@@ -1,18 +1,0 @@
-package com.example.shopping_app;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.fragment.app.Fragment;
-
-public class CheckoutFragment extends Fragment {
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for the Checkout Page
-        return inflater.inflate(R.layout.fragment_checkout, container, false);
-    }
-}
