+++ app/src/main/java/com/example/shopping_app/Fragment/CartFragment.java
@@ -1,4 +1,4 @@
-package com.example.shopping_app;
+package com.example.shopping_app.Fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -13,6 +13,7 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.shopping_app.R;
 import com.google.android.material.button.MaterialButton;
 
 public class CartFragment extends Fragment {
