+++ app/src/main/java/com/example/shopping_app/NotUsedYet/EmptyFavoritesFragment.java
@@ -1,4 +1,4 @@
-package com.example.shopping_app;
+package com.example.shopping_app.NotUsedYet;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -10,11 +10,13 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
+import com.example.shopping_app.Fragment.FavoritesFragment;
+import com.example.shopping_app.R;
 import com.google.android.material.button.MaterialButton;
 
-public class EmptyCartFragment extends Fragment {
+public class EmptyFavoritesFragment extends Fragment {
 
-    public EmptyCartFragment() {
+    public EmptyFavoritesFragment() {
         // Required empty public constructor
     }
 
@@ -23,7 +25,7 @@ public class EmptyCartFragment extends Fragment {
     public View onCreateView(@NonNull LayoutInflater inflater,
                              @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_empty_cart, container, false);
+        return inflater.inflate(R.layout.fragment_empty_favorites, container, false);
     }
 
     @Override
@@ -32,14 +34,12 @@ public class EmptyCartFragment extends Fragment {
         super.onViewCreated(view, savedInstanceState);
 
         ImageButton backButton = view.findViewById(R.id.btn_back);
-        MaterialButton exploreButton = view.findViewById(R.id.btn_explore_categories);
+        MaterialButton exploreButton = view.findViewById(R.id.btn_explore_now);
 
         backButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if (getActivity() != null) {
-                    getActivity().onBackPressed();
-                }
+                if (getActivity() != null) getActivity().onBackPressed();
             }
         });
 
@@ -47,10 +47,10 @@ public class EmptyCartFragment extends Fragment {
             @Override
             public void onClick(View v) {
                 if (getActivity() != null) {
-                    // For demonstration, replace fragment with the cart-with-items fragment
-                    // In a real app, you might navigate to categories or product listings
+                    // Navigate to a fragment or activity where user can explore products
+                    // For demo, we’ll navigate to the FavoritesFragment with items
                     getActivity().getSupportFragmentManager().beginTransaction()
-                            .replace(R.id.cart_screen, new CartFragment())
+                            .replace(R.id.emptyfavorite, new FavoritesFragment())
                             .addToBackStack(null)
                             .commit();
                 }
