+++ app/src/main/java/com/example/shopping_app/NotUsedYet/FavoriteItem.java
@@ -1,4 +1,4 @@
-package com.example.shopping_app;
+package com.example.shopping_app.NotUsedYet;
 
 public class FavoriteItem {
     private String name;
