+++ app/src/main/java/com/example/shopping_app/NotUsedYet/MyItemRecyclerViewAdapter.java
@@ -1,8 +1,7 @@
-package com.example.shopping_app;
+package com.example.shopping_app.NotUsedYet;
 
 import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
-import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
