+++ app/src/main/res/drawable/item_border.xml
@@ -1,7 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <stroke
-        android:width="2dp"
-        android:color="@color/black"
+
+        android:width="1dp"
+        android:color="@color/gray_700"
         />
+    <corners android:radius="40px"></corners>
+
+
 </shape>
