+++ app/src/main/res/layout/favorites_item.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="0dp"
+    android:layout_width="178dp"
     android:layout_height="wrap_content"
     android:layout_margin="8dp"
     android:background="@android:color/white"
