+++ app/src/main/res/layout/fragment_home_screen.xml
@@ -5,7 +5,7 @@
     android:id="@+id/home_screen"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".mainScreen">
+    tools:context=".NotUsedYet.mainScreen">
 
     <LinearLayout
         android:layout_width="match_parent"
