--- app/src/main/res/layout/fragment_product.xml
@@ -1,45 +1,0 @@
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="16dp">
-
-    <LinearLayout
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <!-- Product Image -->
-        <ImageView
-            android:id="@+id/product_image"
-            android:layout_width="match_parent"
-            android:layout_height="300dp"
-            android:scaleType="centerCrop"
-            android:src="@drawable/user_icon_selector" />
-
-        <!-- Product Details -->
-        <TextView
-            android:id="@+id/product_name"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Men's Harrington Jacket"
-            android:textSize="20sp"
-            android:layout_marginTop="16dp" />
-
-        <TextView
-            android:id="@+id/product_price"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="$148"
-            android:textSize="18sp"
-            android:layout_marginTop="8dp" />
-
-        <!-- Add to Cart Button -->
-        <Button
-            android:id="@+id/add_to_cart_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="Add to Cart"
-            android:layout_marginTop="16dp" />
-
-    </LinearLayout>
-</ScrollView>
