+++ app/src/main/res/layout/mainscreen_grid_item.xml
@@ -3,8 +3,10 @@
     android:layout_width="170dp"
     android:layout_height="230dp"
     android:orientation="vertical"
+    android:layout_margin="10dp"
     android:id="@+id/homegriditem"
-    android:background="@drawable/item_border">
+    android:background="@drawable/item_border"
+    >
     <ImageView
         android:id="@+id/ItemImg_home"
     android:layout_width="match_parent"
